requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
asyncio==3.4.3
attrs==23.1.0
black==23.12.1
certifi==2023.11.17
click==8.1.7
dataclasses-json==0.6.3
deepgram-sdk==2.12.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
joblib==1.3.2
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
python-dotenv==1.0.0
regex==2023.12.25
sniffio==1.3.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
verboselogs==1.7
websockets==12.0
yarl==1.9.4
